Imperial Government conversion, bugfixes, and Logging improvements (#38)
* Imperial Government

Countries in I:R which have the imperial government form shall be converted with having CK II's imperial government

* checkFile Method

Created a method to check of a file exists or not. It turns out that it wasn't needed for Born in the Purple calculation, but it could be useful to have anyway

* Imperial Government

* Imperial Government

Added directories needed for imperial government creation

* Create test

* Add files via upload

File used to allow for imperial governments

* Create test

* Add files via upload

File used to allow for imperial governments

* Create test

* Add files via upload

File used to allow for imperial governments

* All imperial governments become empires

I've changed it so that independent countries which have the imperial government form in I:R will automatically be empire tier in CK II, even if they don't meet the size requirements. I've done this because CK II imperial titles cannot create or usurp empire titles, so an imperial government at kingdom tier would be effectively locked to that tier for the rest of the 1,000+ year megacampaign

* All imperial governments become empires

I've changed it so that independent countries which have the imperial government form in I:R will automatically be empire tier in CK II, even if they don't meet the size requirements. I've done this because CK II imperial titles cannot create or usurp empire titles, so an imperial government at kingdom tier would be effectively locked to that tier for the rest of the 1,000+ year megacampaign

* Removed unused scnr variable

* Delete test

* Delete test

* Delete test

* Copy flags, formatting changes

* Copy flags, subjects return overlord's capital, bugfixes

* Rank Parameter in country list

The save file doesn't contain the info, but it's convenient to have as a field to be modified

* checkFile improvements

* Flag directory

* Manichean support

Added mapping for Manichean religion from TinWiz's timeline extension mod

* Kamarupa and Orissa mapping

* New Molon Labe mod formables

Support for new formable nations from the Molon Labe mod

* Create test

* Default placefolder flag

Placefolder flag for countries which don't have any other flag assigned to them

* Delete blank

* Delete test

* Dejure blanking files

Blanked vanilla duchies that were causing problems with the 1066 start date dejure calculation

* c_anglesey typo fix

* Fixed typo

* Added comagoosie

Added credit to comagoosie for developing the Rakaly save decompressor

* Rakaly formatting

Added formatting compatibility for saves decompressed using Rakaly

* Rakaly formatting

Added formatting compatibility for saves decompressed using Rakaly

* File excecution

Added a method to execute the Rakaly decompressor

* Rakaly support

* Rakaly for Windows

Rakaly save decompressor for Windows, created by comagoosie

* Rakaly support

* Typo fix

* Rakaly file

Copies over the rakaly_win.exe file used for Windows save decompression

* Typo fix